requirements.txt
streamlit>=1.28.0
openai>=1.3.0
mistralai>=1.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
PyMuPDF>=1.23.26
Pillow>=10.0.0
supabase>=2.0.0
bcrypt>=4.0.0
